requirements.txt
africastalking==1.1.4
apipkg==1.5
APScheduler==3.5.3
astroid==2.1.0
atomicwrites==1.3.0
attrs==18.2.0
certifi==2018.10.15
chardet==3.0.4
Click==7.0
coverage==4.5.2
coveralls==1.5.1
decorator==4.3.2
dj-database-url==0.5.0
Django==1.11
django-bootstrap3==11.0.0
django-heroku==0.3.1
django-scheduler==0.8.8
djangorestframework==3.6.2
docopt==0.6.2
execnet==1.5.0
gunicorn==19.9.0
icalendar==4.0.3
idna==2.8
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
more-itertools==5.0.0
nose==1.3.7
pathlib2==2.3.3
pep8==1.7.1
pluggy==0.8.1
psycopg2==2.7.5
py==1.7.0
pylint==2.2.2
pytest==4.2.0
pytest-cache==1.0
pytest-cov==2.6.1
python-dateutil==2.7.3
python-decouple==3.1
python-dotenv==0.10.1
python-logstash==0.4.6
pytz==2019.1
pyuploadcare==2.5.0
requests==2.21.0
schedule==0.5.0
schema==0.6.8
six==1.11.0
typed-ast==1.3.0
tzlocal==1.5.1
urllib3==1.24
validators==0.12.4
Werkzeug==0.14.1
whitenoise==4.1
wrapt==1.11.1
